pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clawwrap"
version = "0.1.0"
description = "Spec-first control plane for typed, policy-enforced wrapper contracts"
requires-python = ">=3.11"
dependencies = [
    "PyYAML>=6.0",
    "jsonschema>=4.21",
    "httpx>=0.27",
    "psycopg[binary]>=3.1",
    "psycopg-pool>=3.1",
    "alembic>=1.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.3",
    "mypy>=1.8",
]

[project.scripts]
clawwrap = "clawwrap.cli.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.11"
strict = true

[[tool.mypy.overrides]]
module = "jsonschema"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "jsonschema.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "bandit>=1.9.4",
    "types-jsonschema>=4.26.0.20260202",
]
